Updated zoom.json to work better
Added usage of the, to be released, JWT authentication
Added usage of the, to be released, paging options
It now uses totalPageAttribute, but totalAttribute is also supported, but zoom somehow does not return as much records as they say they do, so this does not work correctly for zoom. In both attributes u can specify a full JSONPath, but $. is always appended, so u can provide a complete JSON path starting from the root. Like 
attributeName.child[4].somethingInsideAnArray should work.